AppAgentX/backend/ImageEmbedding/requirements.txt
fastapi==0.110.0
uvicorn==0.27.0
torch==2.2.0
timm==0.9.12
pillow==10.2.0
torchvision==0.17.0
pydantic==2.6.0
python-multipart==0.0.9
numpy==1.26.0 